Change the type of the four points to int

+++ modules/serial/packet.h
@@ -35,10 +35,10 @@ struct SerialSendPacket {
     float delay;
     int distance_mode;  // maybe only used in sentry
     int fire;
-    short point1_x, point1_y;
-    short point2_x, point2_y;
-    short point3_x, point3_y;
-    short point4_x, point4_y;
+    int point1_x, point1_y;
+    int point2_x, point2_y;
+    int point3_x, point3_y;
+    int point4_x, point4_y;
     float check_sum;
 };
